Financial_API/requirements.txt
fastapi
uvicorn
sqlalchemy  
psycopg2-binary
python-dotenv  
bcrypt
email-validator
python-jose

